backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
PyJWT==2.8.0
firebase-admin==6.0.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
